feat(urlFavorites): implement URL favorites management module with hooks and components
- Add UrlFavoritesStatus component for displaying status messages.
- Create UrlFavoritesTab component for managing and editing URL favorites in expanded mode.
- Define constants and configuration for URL favorites in urlFavoritesConfig.ts.
- Implement useUrlFavorites hook for state management and logic handling.
- Create index.ts for unified exports of the URL favorites module.
- Define types for URL favorites in urlFavoritesTypes.ts.
- Add utility functions for parsing, serializing, and managing URL favorites in urlFavoritesUtils.ts.
