+++ src/renderer/components/components/StandaloneWindowViewport.tsx
@@ -27,7 +27,7 @@
 import type { JSX } from 'react';
 import { TodoTab } from '../states/maxExpand/components/todo/components/TodoTab';
 import { CountdownTab } from '../states/maxExpand/components/countdown';
-import { UrlFavoritesTab } from '../states/maxExpand/components/UrlFavoritesTab';
+import { UrlFavoritesTab } from '../states/maxExpand/components/urlFavorites';
 import { AlbumTab } from '../states/maxExpand/components/album/components/AlbumTab';
 import { MailTab } from '../states/maxExpand/components/mail';
 import { LocalFileSearchTab } from '../states/maxExpand/components/localFileSearch/components/LocalFileSearchTab';
