+++ src/renderer/components/states/maxExpand/MaxExpandContentEager.tsx
@@ -29,7 +29,7 @@ import type { MaxExpandTab } from '../../../store/types';
 import { MaxExpandContentShell } from './MaxExpandContentShell';
 import { AiChatTab } from './components/agent/components/AiChatTab';
 import { TodoTab } from './components/todo/components/TodoTab';
-import { UrlFavoritesTab } from './components/UrlFavoritesTab';
+import { UrlFavoritesTab } from './components/urlFavorites';
 import { LocalFileSearchTab } from './components/localFileSearch/components/LocalFileSearchTab';
 import { ClipboardHistoryTab } from './components/clipBoardHistory';
 import { AlbumTab } from './components/album/components/AlbumTab';
